# openapi/api.yaml
openapi: 3.0.3
info:
  title: RevU API
  description: RevU API for businesses and customers.
  version: 1.0.2
servers:
  - url: 'http://34.209.63.8:5000/api/v1'
    description: Production
  - url: 'http://localhost:5000/api/v1'
    description: Development
security:
  - bearerAuth: []
paths:
  /auth/login:
    post:
      summary: Login a user
      tags:
        - auth
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  $ref: '#/components/schemas/Role'
                mobile:
                  $ref: '#/components/schemas/Mobile'
                password:
                  $ref: '#/components/schemas/MediumText'
                otp:
                  $ref: '#/components/schemas/OTP'
              required:
                - role
                - mobile
            examples:
              passwordLogin:
                $ref: '#/components/examples/passwordLogin'
              otpLogin:
                $ref: '#/components/examples/otpLogin'
      responses:
        '201':
          description: Logged in successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthToken'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /auth/refresh:
    get:
      summary: Refresh token
      tags:
        - auth
      operationId: refreshToken
      responses:
        '200':
          description: New access token issued successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Token'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /appointments:
    post:
      summary: Create an appointment
      tags:
        - appointments
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
            examples:
              appointment:
                $ref: '#/components/examples/appointment'
      responses:
        '201':
          description: Appointment created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get appointments list
      tags:
        - appointments
      operationId: getAppointments
      parameters:
        - $ref: '#/components/parameters/fromDate'
        - $ref: '#/components/parameters/toDate'
        - $ref: '#/components/parameters/appointmentStatus'
        - $ref: '#/components/parameters/salonId'
        - $ref: '#/components/parameters/employeeId'
        - $ref: '#/components/parameters/appointmentIncludeWithPagination'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
      responses:
        '200':
          description: Paginated appointments list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedListSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      summary: Delete appointments
      tags:
        - appointments
      operationId: deleteAppointments
      parameters:
        - $ref: '#/components/parameters/deleteIds'
      responses:
        '200':
          description: Appointments deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiObjectOperationSuccessResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /appointments/{id}:
    get:
      summary: Get an appointment
      tags:
        - appointments
      operationId: getAppointment
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/appointmentInclude'
      responses:
        '200':
          description: Appointment details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    patch:
      summary: Update an appointment
      tags:
        - appointments
      operationId: updateAppointment
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentUpdateBody'
            examples:
              appointment:
                $ref: '#/components/examples/appointment'
      responses:
        '201':
          description: Appointment created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Appointment'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /coupons:
    post:
      summary: Create a coupon
      tags:
        - coupons
      operationId: createCoupon
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Coupon'
            examples:
              coupon:
                $ref: '#/components/examples/coupon'
      responses:
        '201':
          description: Coupon created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Coupon'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    get:
      summary: Get coupons list
      tags:
        - coupons
      operationId: getCoupons
      parameters:
        - $ref: '#/components/parameters/couponCode'
        - $ref: '#/components/parameters/customerId'
        - $ref: '#/components/parameters/salonId'
        - $ref: '#/components/parameters/couponIncludeWithPagination'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/perPage'
      responses:
        '200':
          description: Paginated coupons list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedListSuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Coupon'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    delete:
      summary: Delete coupons
      tags:
        - coupons
      operationId: deleteCoupons
      parameters:
        - $ref: '#/components/parameters/deleteIds'
      responses:
        '200':
          description: Coupons deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiObjectOperationSuccessResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /coupons/{id}:
    get:
      summary: Get a coupon
      tags:
        - coupons
      operationId: getCoupon
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/couponInclude'
      responses:
        '200':
          description: Coupon details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Coupon'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
    patch:
      summary: Update a coupon
      tags:
        - coupons
      operationId: updateCoupon
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CouponUpdateBody'
            examples:
              coupon:
                $ref: '#/components/examples/coupon'
      responses:
        '201':
          description: Coupon updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SingleSuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Coupon'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
components:
  parameters:
    id:
      name: id
      description: Resource id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/IntegerId'
...
    deleteIds:
      name: ids
      description: List of ids to delete
      in: query
      required: true
      style: form
      explode: false
      schema:
        $ref: '#/components/schemas/IntegerIds'
...
    fromDate:
      name: fromDate
      description: Limit results after date
      in: query
      schema:
        $ref: '#/components/schemas/DateTime'
    toDate:
      name: toDate
      description: Limit results before date
      in: query
      schema:
        $ref: '#/components/schemas/DateTime'
    salonId:
      name: salonId
      description: Limit results to specific salon id
      in: query
      schema:
        $ref: '#/components/schemas/IntegerId'
...
    appointmentStatus:
      name: status
      description: Limit results with specific status
      in: query
      schema:
        $ref: '#/components/schemas/AppointmentStatus'
    employeeId:
      name: employeeId
      description: Limit results to specific employee id
      in: query
      schema:
        $ref: '#/components/schemas/IntegerId'
...
    customerId:
      name: customerId
      description: Limit results to specific customer id
      in: query
      schema:
        $ref: '#/components/schemas/IntegerId'
...
    couponCode:
      name: code
      description: Limit results matching coupon code
      in: query
      schema:
        $ref: '#/components/schemas/ShortText'
    appointmentInclude:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - appointmentSlots
    appointmentIncludeWithPagination:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - appointmentSlots
            - _pagination.total
...
    couponInclude:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - restrictions
    couponIncludeWithPagination:
      name: include
      description: Get extra fields in result
      in: query
      style: form
      explode: false
      schema:
        type: array
        items:
          enum:
            - restrictions
            - _pagination.total
...
    page:
      name: page
      description: Current page no. for pagination
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    perPage:
      name: perPage
      description: No. of records in paginated list
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
...
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    '500':
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    PositiveIntegerIncludingZero:
      type: integer
      minimum: 0
    IntegerId:
      type: integer
      minimum: 1
    IntegerIds:
      type: array
      items:
        $ref: '#/components/schemas/IntegerId'
      minItems: 1
    ShortText:
      type: string
      minLength: 2
      maxLength: 64
    MediumText:
      type: string
      minLength: 2
      maxLength: 255
...
    Money:
      type: number
      format: float
      minimum: 0
    Mobile:
      type: string
      format: regex
      pattern: '^[6-9]\d{9}$'
    OTP:
      type: string
      format: regex
      pattern: '^\d{6}$'
    DateTime:
      type: string
      format: date-time
      description: ISO 8601 compliant date time
...
    Role:
      enum:
        - customer
        - merchant
        - employee
        - admin
...
    RestrictionType:
      enum:
        - only
        - include
        - exclude
...
    DiscountType:
      enum:
        - flat
        - percent
...
    EntityStatus:
      enum:
        - created
        - verified
        - disabled
        - trashed
...
    Token:
      type: object
      properties:
        token:
          type: string
          minLength: 64
        expiresAt:
          $ref: '#/components/schemas/DateTime'
      required:
        - token
    AuthToken:
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/IntegerId'
        roleUserId:
          $ref: '#/components/schemas/IntegerId'
        role:
          $ref: '#/components/schemas/Role'
        accessToken:
          $ref: '#/components/schemas/Token'
        refreshToken:
          $ref: '#/components/schemas/Token'
...
    User:
      allOf:
        - $ref: '#/components/schemas/UserUpdateBody'
      required:
        - firstName
        - gender
        - mobile
...
    AppointmentStatus:
      enum:
        - upcoming
        - ongoing
        - completed
        - cancelled
        - rejected
    AppointmentUpdateBody:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        appointmentId:
          $ref: '#/components/schemas/ShortText'
        customer:
          allOf:
            - $ref: '#/components/schemas/Customer'
          readOnly: true
        customerId:
          $ref: '#/components/schemas/IntegerId'
        salonId:
          $ref: '#/components/schemas/IntegerId'
        startsAt:
          $ref: '#/components/schemas/DateTime'
        endsAt:
          $ref: '#/components/schemas/DateTime'
        appointmentSlots:
          type: array
          items:
            $ref: '#/components/schemas/AppointmentSlot'
          minItems: 1
        otp:
          allOf:
            - $ref: '#/components/schemas/OTP'
          readOnly: true
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        createdAt:
          $ref: '#/components/schemas/DateTime'
    Appointment:
      allOf:
        - $ref: '#/components/schemas/AppointmentUpdateBody'
      required:
        - startsAt
        - endsAt
        - customerId
        - salonId
    AppointmentSlot:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        appointmentId:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        optedServiceId:
          $ref: '#/components/schemas/IntegerId'
        employeeId:
          $ref: '#/components/schemas/IntegerId'
        startsAt:
          $ref: '#/components/schemas/DateTime'
        endsAt:
          $ref: '#/components/schemas/DateTime'
        note:
          type: string
      required:
        - appointmentId
        - optedServiceId
...
    CouponUpdateBody:
      type: object
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/IntegerId'
          readOnly: true
        code:
          $ref: '#/components/schemas/ShortText'
        description:
          type: string
        discountType:
          $ref: '#/components/schemas/DiscountType'
        amount:
          $ref: '#/components/schemas/Money'
        minPurchaseAmount:
          $ref: '#/components/schemas/Money'
        maxPurchaseAmount:
          $ref: '#/components/schemas/Money'
        individualUse:
          type: boolean
        usageLimit:
          type: integer
          minimum: 1
        usageCount:
          type: integer
          minimum: 0
          readOnly: true
        expiresAt:
          type: string
          format: date-time
        restrictions:
          type: array
          items:
            $ref: '#/components/schemas/Restriction'
          uniqueItems: true
        status:
          $ref: '#/components/schemas/EntityStatus'
        createdAt:
          type: string
          format: date-time
    Coupon:
      allOf:
        - $ref: '#/components/schemas/CouponUpdateBody'
      required:
        - code
        - discountType
        - amount
    Restriction:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/RestrictionType'
        customerId:
          $ref: '#/components/schemas/IntegerId'
        salonId:
          $ref: '#/components/schemas/IntegerId'
        employeeId:
          $ref: '#/components/schemas/IntegerId'
        serviceId:
          $ref: '#/components/schemas/IntegerId'
        categoryId:
          $ref: '#/components/schemas/IntegerId'
        couponId:
          $ref: '#/components/schemas/IntegerId'
      required:
        - type
    CustomerUpdateBody:
      allOf:
        - $ref: '#/components/schemas/UserUpdateBody'
        - type: object
          properties:
            userId:
              allOf:
                - $ref: '#/components/schemas/IntegerId'
              readOnly: true
    Customer:
      allOf:
        - $ref: '#/components/schemas/User'
        - $ref: '#/components/schemas/CustomerUpdateBody'
...
    Error:
      type: object
      properties:
        code:
          oneOf:
            - type: number
            - type: string
        message:
          type: string
        data:
          oneOf:
            - type: object
            - type: array
              items: {}
      required:
        - message
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        perPage:
          type: integer
          minimum: 1
        total:
          type: integer
      required:
        - page
        - perPage
    JSONResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/JSONResponse'
        - type: object
          properties:
            success:
              type: boolean
              default: true
    SingleSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
          required:
            - data
    ListSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: array
              items: {}
          required:
            - data
    PaginatedListSuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ListSuccessResponse'
        - type: object
          properties:
            _pagination:
              $ref: '#/components/schemas/Pagination'
          required:
            - _pagination
...
    MultiObjectOperationSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
        data:
          type: object
          properties:
            count:
              $ref: '#/components/schemas/PositiveIntegerIncludingZero'
          required:
            - count
      required:
        - success
        - data
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/JSONResponse'
        - type: object
          properties:
            error:
              $ref: '#/components/schemas/Error'
          required:
            - error
  examples:
...
    passwordLogin:
      value:
        role: 'merchant'
        mobile: '9862898720'
        password: 'pa$$word'
    otpLogin:
      value:
        role: 'customer'
        mobile: '9876543210'
        otp: '123456'
...
    coupon:
      value:
        code: OFF30
        description: 30% off on min spend value of 2000.
        discountType: percent
        amount: 30
        minPurchaseAmount: 2000
        individualUse: true
        usageLimit: 20
        expiresAt: '2022-09-15T19:26:34.769Z'
        restrictions:
          - type: include
            customerId: 1
          - type: include
            customerId: 2
...
    appointment:
      value:
        appointmentId: AP23423432233
        customerId: 1
        salonId: 1
        startsAt: '2022-08-15T19:26:34.769Z'
        endsAt: '2022-08-15T22:26:34.739Z'
        appointmentSlots:
          - optedServiceId: 1
            employeeId: 1
            startsAt: '2022-08-15T19:26:34.769Z'
            endsAt: '2022-08-15T20:26:34.769Z'
            note: 'Some note'
          - optedServiceId: 2
            employeeId: 1
            startsAt: '2022-08-15T21:26:34.769Z'
            endsAt: '2022-08-15T22:26:34.769Z'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT